Moved Project model to separate app
