--- FOSSDB_web/apps/fossdb/forms.py
@@ -1,22 +1,0 @@
-from django import forms
-
-from .models import Project
-
-
-class ProjectForm(forms.ModelForm):
-
-    class Meta:
-        model = Project
-        fields = ["name", "description", "licenses"]
-
-        widgets = {
-            "name": forms.TextInput(attrs={
-                "class": "form-control",
-                "placeholder": "Project name",
-            }),
-            "description": forms.Textarea(attrs={
-                "class": "form-control",
-                "placeholder": "Description",
-            }),
-            "licenses": forms.CheckboxSelectMultiple(),
-        }
