+++ FOSSDB_web/apps/fossdb/hosting_platform/forms.py
@@ -1,0 +1,11 @@
+from django import forms
+
+from .models import HostingPlatform
+
+
+class HostingPlatformForm(forms.ModelForm):
+    url = forms.URLField()
+
+    class Meta:
+        model = HostingPlatform
+        fields = ["hosting_platform", "url"]
