+++ FOSSDB_web/apps/fossdb/programming_language/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ProgrammingLanguageConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "fossdb.programming_language"
