+++ FOSSDB_web/apps/fossdb/project/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ProjectConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "fossdb.project"
