+++ FOSSDB_web/apps/fossdb/tag/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class TagConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "fossdb.tag"
