--- FOSSDB_web/apps/tag/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class TagConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'tag'
